backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.30,<0.31
python-multipart>=0.0.9,<0.1
pydantic>=2.7,<3
pydantic-settings>=2.3,<3
sqlmodel==0.0.21
passlib[bcrypt]>=1.7,<2
python-jose[cryptography]>=3.3,<4
APScheduler>=3.10,<3.11
python-telegram-bot>=21.6,<22
alembic>=1.13,<2
